+++ compiler/__attribute__/no_instrument_function.c
@@ -1,0 +1,1 @@
+../../ftrace/samples/mcount-nop.c