+++ CHANGELOG.md
@@ -49,8 +49,9 @@ and this project adheres to
 - hpc: add gpu/abbrev.md;
 - kunlunxin: add info.sh;
 - hpc/kunlunxin: Introduce;
-- ftrace: Move ftrace/demos to ftrace/samples;
 #### Changed
+- ftrace: Move mcount-nop.c to ftrace/samples/;
+- ftrace: Move ftrace/demos to ftrace/samples;
 - targets/nvidia.mk: split some code to cuda.mk;
 - cuda_adapter.h: Rename macros __pfx and __PFX to __cu,__CU;
 - linker: Rename linker-script to script;